requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
httpx>=0.25.0
aiofiles>=23.0.0
openai>=1.12.0
anthropic>=0.18.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
pytest>=7.4.0
hyperliquid-python-sdk>=0.4.0
eth-account>=0.10.0
